Rename advnace to advanced
