+++ openmetadata-ui/src/main/resources/ui/src/utils/AdvancedSearchUtils.ts
